Move debug overlay to a OpenSage.Game.Gui.DebugUi
